requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2
django-ckeditor==6.2.0
django-js-asset==1.2.2
Pillow==9.0.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
